firearm_analysis_package/requirements.txt
plaintext
pandas
matplotlib
folium
selenium
Pillow
unittest
